further optimizations to langQuiz
